Removed dependency on Typesafe releases repo
Fixes #3436.

* Upgraded to netty-http-pipelining 1.1.4 (which is now deployed to
  maven central)
* Removed Typesafe releases repo resolvers from play settings

I tested this locally with a clean ivy cache, and found an additional
issue that specs2 now depends on scalaz-stream as a required dependency,
so a note was added to the migration guide and the play scala template
has been updated.

Note that this only removes the dependency on the Typesafe releases repo
for compile/runtime dependencies.  For build dependencies (sbt plugins
etc) this is being tracked in #4223.