Attempted build stability improvements
This attempts to fix two build stability problems. One is an OOME from
scalac when compiling tests. Hard to know what's causing this, but we
know that scalac has problems compiling the same source code twice in
the same JVM, these processes are isolated in the Compile scope due to a
specific bug with macros, but maybe there are more bugs, so isolating
them so the run sequentially in the Test scope may work.

The other is random failures in the dev mode scripted test, and I can't
see what the problem is. I'm inserting a sleep at an opportune time,
not because I think it should be there, but because I want to know if
that makes a difference.