Split play-jdbc into three different modules to provide different `DBApi` impl
Split play-jdbc into three modules:

* play-jdbc-api: Containing only the `Database` and `DBApi` traits, for both
		 Java and Scala.
* play-jdbc-evolutions: Evolutions support, depending only on play-jdbc-api.
* play-jdbc: Play JDBC support, depending only on play-jdbc-api. In fact, JDBC
	     is now completely decoupled from evolutions. To use evolutions
             users are required to explicitly add the `evolutions` components to their
             project's dependencies.

The modularization allows other modules (in the specific case, play-slick) to
provide a binding for `DBApi` so that evolutions can be supported with minimum
work.

Another benefit is that all the play-jdbc classes, which are not needed when
using play-slick, are not in the application's classpath, preventing any
misusage.

Note that play-slick is only an example that motivates why the modularization is
worth. There is nothing specific to play-slick in this commit.

A final remark. I've thought about splitting both play-jdbc-api and
play-jdbc-evolutions even further, so that Scala user's would not see in their
classpath any Java class. Personally, I think this is a refactoring that is
worth doing, as it improves the developer experience for our Scala users.
However, before doing it, I'd like to make sure this is something you would
support.